more confused

+++ matrix_vector_mul.c
@@ -2,9 +2,9 @@
 #include <stdlib.h>
 #include <time.h>
 
-void matrix_vector_product(float *a, float *v, float *res, int matrix_size) {
+void matrix_vector_product(long double *a, long double *v, long double *res, int matrix_size) {
     for (int i = 0; i < matrix_size; i++) {
-        float sum = 0.0;
+        long double sum = 0.0;
         for (int j = 0; j < matrix_size; j++) {
             sum += a[i * matrix_size + j] * v[j];
         } 
@@ -13,18 +13,18 @@ void matrix_vector_product(float *a, float *v, float *res, int matrix_size) {
 }
 
 int main(int argc, char **argv) {
-    float *a, *v, *res; 
+    long double *a, *v, *res; 
     int matrix_size = atoi(argv[1]);
     
-    a = (float *) malloc(matrix_size * matrix_size * sizeof(float));
-    v = (float *) malloc(matrix_size * sizeof(float));
-    res = (float *) malloc(matrix_size * sizeof(float));
+    a = (long double *) malloc(matrix_size * matrix_size * sizeof(long double));
+    v = (long double *) malloc(matrix_size * sizeof(long double));
+    res = (long double *) malloc(matrix_size * sizeof(long double));
     
     for (int i = 0; i < matrix_size; i++) {
         for (int j = 0; j < matrix_size; j++) {
-            a[i * matrix_size + j] = (float) i * matrix_size + j;
+            a[i * matrix_size + j] = (long double) i * matrix_size + j;
         }
-        v[i] = (float) i;
+        v[i] = (long double) i;
     }
 
     clock_t st, en;
@@ -33,7 +33,7 @@ int main(int argc, char **argv) {
     en = clock();
 
     for (int i = 0; i < matrix_size; i++) {
-        printf("%.2f\n", res[i]);
+        printf("%.2Lf\n", res[i]);
     }
     
     printf("Time taken by CPU: %f seconds\n", (double)(en - st) / CLOCKS_PER_SEC);
